Move `RING_OF_LIGHT` to its own file
