+++ scripts/modules/itemacros/items.mjs
@@ -1,6 +1,7 @@
 import { DEPEND } from "../../const.mjs";
 import { ItemMacroHelpers } from "../itemMacros.mjs";
 import { HIT_DIE_APPLY } from "./items/hitDieApply.mjs";
+import { RING_OF_LIGHT } from "./items/ringOfLight.mjs";
 import { TORCH } from "./items/torch.mjs";
 
 export const items = {
@@ -10,21 +11,6 @@ export const items = {
   RING_OF_LIGHT,
   TORCH,
 };
-
-async function RING_OF_LIGHT(item, speaker, actor, token, character, event, args) {
-  if (!ItemMacroHelpers._getDependencies(DEPEND.EM, DEPEND.VAE)) return item.use();
-
-  if (!item.system.equipped) await item.update({ "system.equipped": true }); // Forcefully equip the item.
-
-  const has = actor.effects.find((e) => e.statuses.has(item.name.slugify({ strict: true })));
-  if (has) return has.delete();
-
-  const use = await item.use();
-  if (!use) return;
-
-  const lightData = { bright: 30, dim: 60 };
-  return actor.createEmbeddedDocuments("ActiveEffect", ItemMacroHelpers._constructLightEffectData({ item, lightData }));
-}
 
 async function LANTERN_OF_TRACKING(item, speaker, actor, token, character, event, args) {
   if (!ItemMacroHelpers._getDependencies(DEPEND.EM, DEPEND.VAE)) return item.use();
